Install check script

+++ orchestration/run_container/bind.py
@@ -5,6 +5,13 @@ class Bind(Container):
     def update(self, config_timestamp):
         with open(f"output/{config_timestamp}/etc-bind.tar", "rb") as f:
             self.container.put_archive("/etc/bind", f.read())
+
+        # call named checks
+        (exit_code, output) = self.container.exec_run('/etc/bind/named-check.sh')
+        self.logger.info(output.decode())
+        if exit_code != 0:
+            raise Exception("named-check.sh failed")
+
         self.container.kill(signal="SIGHUP")
 
     def toggle_recursion(self, recursion):
